Applied the general lobby to ranked as well
