+++ src/Frontend/css/matchCenter/tailwindOutput.css
@@ -380,6 +380,10 @@
     border-style: var(--tw-border-style);
     border-width: 2px;
   }
+  .border-b {
+    border-bottom-style: var(--tw-border-style);
+    border-bottom-width: 1px;
+  }
   .border-l {
     border-left-style: var(--tw-border-style);
     border-left-width: 1px;
@@ -395,6 +399,12 @@
   }
   .border-gray-900 {
     border-color: var(--color-gray-900);
+  }
+  .border-gray-900\/50 {
+    border-color: color-mix(in srgb, oklch(21% 0.034 264.665) 50%, transparent);
+    @supports (color: color-mix(in lab, red, red)) {
+      border-color: color-mix(in oklab, var(--color-gray-900) 50%, transparent);
+    }
   }
   .border-gray-900\/75 {
     border-color: color-mix(in srgb, oklch(21% 0.034 264.665) 75%, transparent);
